requirements.txt
Flask>=2.0
flask-cors
pandas
numpy
scikit-learn
xgboost
joblib
tensorflow
requests
python-dotenv